requirements.txt
streamlit==1.45.0
requests==2.32.3
python-dotenv>=1.0.0